test(sam): split provider transaction coverage
